Support top-level let in a libarary;
the top-level code gets compiled to an exported main function. The
main function compilation reuses the compile-procedure-body procedure
from the expression-compiler to process any local definitions
generated into the body program by let expressions.

Extend semantics of local lexical frames by defining an offset of zero
for a local frame that follows the global lexical frame. This needs to
be done because the compiler allocates the global lexical frame's
variables from the global index space, which is is separate from the
local index space.